+++ RELEASE_NOTES.md
@@ -1,6 +1,12 @@
 Release notes
 -------------
 
+## WebTex 1.3.0
+
+* Added X-MathImage-width and X-MathImage-height headers. This is a 
+  backwards compatible API change adding to the previously sent headers.
+* Use documentclass standalone to avoid too narrow typesetting.
+
 ## WebTex 1.2.2
 
 * The X-MathImage-tex header should be encoded.